Create 5 rounds of rock, paper, scissors
By declaring the score variables and playRound() function inside the
scope of the playGame() function, and running the playRound() function
in a for loop, it makes accessing code more orderly as it is all in
one place.

Creating 5 rounds also allows the game to run for a bit longer too.

Keep